% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{prediccion}
\alias{prediccion}
\title{Predicciones de la regresión polinomial}
\usage{
prediccion(x, betas)
}
\arguments{
\item{x}{(vector) Datos de la variable independiente (explicativa)}

\item{betas}{(vector) Coeficientes del modelo polinomial}
}
\value{
Una matriz con los valores de Y calculados
}
\description{
Realiza predicciones a partir de un modelo dado
}
\examples{
\dontrun
#Directorio de trabajo
ruta<- "~/ruta/datos.csv"
(datos<- read.csv(ruta))

#Cargando libreria
library(ReP)

b<- regpol(x=datos$x, y=datos$y, 4)
beta<-b[[1]][1]
nuevaX<-c(5,6,7,8,9,10)

prediccion(nuevaX, beta)
}
