% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPF.R
\name{regpol}
\alias{regpol}
\title{Regresión Polinomial}
\usage{
regpol(x, y, grado)
}
\arguments{
\item{x}{(vector) Datos de la variable independiente (explicativa)}

\item{y}{(vector) Datos de la variable dependiente (respuesta)}

\item{grado}{(valor) Grado al que se desea realizar el polinomio}
}
\value{
Una lista con los valores de Y ajustada, los coeficientes de regresión,el error estandar, la R^2 y la varianza
}
\description{
Realiza una regresión polinomial a partir del grado indicado
}
\examples{
\dontrun
#Cargando la libreria
library(ReP)

Datos para el modelo
X<-(1, 3, 5, 7, 9)
Y<-(4, 8, 12, 16, 20)
salida <-  regpol(x=X,y=Y, grado=3)

}
